// amplify-migration-apps/project-boards-datastore/src/API.ts
export enum ProjectStatus {
  ACTIVE = 'ACTIVE',
  COMPLETED = 'COMPLETED',
  ON_HOLD = 'ON_HOLD',
  ARCHIVED = 'ARCHIVED',
}
// ...
export type ModelStringInput = {
  ne?: string | null;
  eq?: string | null;
  le?: string | null;
  lt?: string | null;
  ge?: string | null;
  gt?: string | null;
  contains?: string | null;
  notContains?: string | null;
  between?: Array<string | null> | null;
  beginsWith?: string | null;
  attributeExists?: boolean | null;
  attributeType?: ModelAttributeTypes | null;
  size?: ModelSizeInput | null;
};
// ...
export type ModelProjectStatusInput = {
  eq?: ProjectStatus | null;
  ne?: ProjectStatus | null;
};

export type ModelBooleanInput = {
  ne?: boolean | null;
  eq?: boolean | null;
  attributeExists?: boolean | null;
  attributeType?: ModelAttributeTypes | null;
};

export type Project = {
  __typename: 'Project';
  id: string;
  title: string;
  description?: string | null;
  status: ProjectStatus;
  deadline?: string | null;
  color?: string | null;
  todos?: ModelTodoConnection | null;
  createdAt: string;
  updatedAt: string;
  _version: number;
  _deleted?: boolean | null;
  _lastChangedAt: number;
};

export type ModelTodoConnection = {
  __typename: 'ModelTodoConnection';
  items: Array<Todo | null>;
  nextToken?: string | null;
  startedAt?: number | null;
};

export type Todo = {
  __typename: 'Todo';
  id: string;
  name: string;
  description?: string | null;
  images?: Array<string | null> | null;
  projectID?: string | null;
  createdAt: string;
  updatedAt: string;
  _version: number;
  _deleted?: boolean | null;
  _lastChangedAt: number;
  projectTodosId?: string | null;
};
// ...
export type ModelIDInput = {
  ne?: string | null;
  eq?: string | null;
  le?: string | null;
  lt?: string | null;
  ge?: string | null;
  gt?: string | null;
  contains?: string | null;
  notContains?: string | null;
  between?: Array<string | null> | null;
  beginsWith?: string | null;
  attributeExists?: boolean | null;
  attributeType?: ModelAttributeTypes | null;
  size?: ModelSizeInput | null;
};

export type UpdateTodoInput = {
  id: string;
  name?: string | null;
  description?: string | null;
  images?: Array<string | null> | null;
  projectID?: string | null;
  _version?: number | null;
  projectTodosId?: string | null;
};

export type DeleteTodoInput = {
  id: string;
  _version?: number | null;
};

export type QuoteResponse = {
  __typename: 'QuoteResponse';
  message: string;
  quote: string;
  author: string;
  timestamp: string;
  totalQuotes: number;
};

export type ModelProjectFilterInput = {
  id?: ModelIDInput | null;
  title?: ModelStringInput | null;
  description?: ModelStringInput | null;
  status?: ModelProjectStatusInput | null;
  deadline?: ModelStringInput | null;
  color?: ModelStringInput | null;
  createdAt?: ModelStringInput | null;
  updatedAt?: ModelStringInput | null;
  and?: Array<ModelProjectFilterInput | null> | null;
  or?: Array<ModelProjectFilterInput | null> | null;
  not?: ModelProjectFilterInput | null;
  _deleted?: ModelBooleanInput | null;
};

export type ModelProjectConnection = {
  __typename: 'ModelProjectConnection';
  items: Array<Project | null>;
  nextToken?: string | null;
  startedAt?: number | null;
};

export type ModelTodoFilterInput = {
  id?: ModelIDInput | null;
  name?: ModelStringInput | null;
  description?: ModelStringInput | null;
  images?: ModelStringInput | null;
  projectID?: ModelIDInput | null;
  createdAt?: ModelStringInput | null;
  updatedAt?: ModelStringInput | null;
  and?: Array<ModelTodoFilterInput | null> | null;
  or?: Array<ModelTodoFilterInput | null> | null;
  not?: ModelTodoFilterInput | null;
  _deleted?: ModelBooleanInput | null;
  projectTodosId?: ModelIDInput | null;
};
// ...
export type GetRandomQuoteQueryVariables = {};

export type GetRandomQuoteQuery = {
  getRandomQuote?: {
    __typename: 'QuoteResponse';
    message: string;
    quote: string;
    author: string;
    timestamp: string;
    totalQuotes: number;
  } | null;
};

export type GetProjectQueryVariables = {
  id: string;
};

export type GetProjectQuery = {
  getProject?: {
    __typename: 'Project';
    id: string;
    title: string;
    description?: string | null;
    status: ProjectStatus;
    deadline?: string | null;
    color?: string | null;
    todos?: {
      __typename: 'ModelTodoConnection';
      nextToken?: string | null;
      startedAt?: number | null;
    } | null;
    createdAt: string;
    updatedAt: string;
    _version: number;
    _deleted?: boolean | null;
    _lastChangedAt: number;
  } | null;
};

export type ListProjectsQueryVariables = {
  filter?: ModelProjectFilterInput | null;
  limit?: number | null;
  nextToken?: string | null;
};

export type ListProjectsQuery = {
  listProjects?: {
    __typename: 'ModelProjectConnection';
    items: Array<{
      __typename: 'Project';
      id: string;
      title: string;
      description?: string | null;
      status: ProjectStatus;
      deadline?: string | null;
      color?: string | null;
      createdAt: string;
      updatedAt: string;
      _version: number;
      _deleted?: boolean | null;
      _lastChangedAt: number;
    } | null>;
    nextToken?: string | null;
    startedAt?: number | null;
  } | null;
};

export type SyncProjectsQueryVariables = {
  filter?: ModelProjectFilterInput | null;
  limit?: number | null;
  nextToken?: string | null;
  lastSync?: number | null;
};

export type SyncProjectsQuery = {
  syncProjects?: {
    __typename: 'ModelProjectConnection';
    items: Array<{
      __typename: 'Project';
      id: string;
      title: string;
      description?: string | null;
      status: ProjectStatus;
      deadline?: string | null;
      color?: string | null;
      createdAt: string;
      updatedAt: string;
      _version: number;
      _deleted?: boolean | null;
      _lastChangedAt: number;
    } | null>;
    nextToken?: string | null;
    startedAt?: number | null;
  } | null;
};

export type GetTodoQueryVariables = {
  id: string;
};

export type GetTodoQuery = {
  getTodo?: {
    __typename: 'Todo';
    id: string;
    name: string;
    description?: string | null;
    images?: Array<string | null> | null;
    projectID?: string | null;
    createdAt: string;
    updatedAt: string;
    _version: number;
    _deleted?: boolean | null;
    _lastChangedAt: number;
    projectTodosId?: string | null;
  } | null;
};

export type ListTodosQueryVariables = {
  filter?: ModelTodoFilterInput | null;
  limit?: number | null;
  nextToken?: string | null;
};

export type ListTodosQuery = {
  listTodos?: {
    __typename: 'ModelTodoConnection';
    items: Array<{
      __typename: 'Todo';
      id: string;
      name: string;
      description?: string | null;
      images?: Array<string | null> | null;
      projectID?: string | null;
      createdAt: string;
      updatedAt: string;
      _version: number;
      _deleted?: boolean | null;
      _lastChangedAt: number;
      projectTodosId?: string | null;
    } | null>;
    nextToken?: string | null;
    startedAt?: number | null;
  } | null;
};

export type SyncTodosQueryVariables = {
  filter?: ModelTodoFilterInput | null;
  limit?: number | null;
  nextToken?: string | null;
  lastSync?: number | null;
};

export type SyncTodosQuery = {
  syncTodos?: {
    __typename: 'ModelTodoConnection';
    items: Array<{
      __typename: 'Todo';
      id: string;
      name: string;
      description?: string | null;
      images?: Array<string | null> | null;
      projectID?: string | null;
      createdAt: string;
      updatedAt: string;
      _version: number;
      _deleted?: boolean | null;
      _lastChangedAt: number;
      projectTodosId?: string | null;
    } | null>;
    nextToken?: string | null;
    startedAt?: number | null;
  } | null;
};